pyproject.toml
[tool.poetry]
name = "interview-assessment"
version = "1.0.0"
description = "Interview Assessment"
authors = ["Engineering Team <engineering.team@progyny.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
PyMySQL = "^1.0.2"
cryptography = "^37.0.3"
sqlalchemy = "^1.4.39"
apscheduler = "^3.9.1"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.20.0"
pytest = "^7.1.2"
pytest-mock = "^3.8.1"
pytest-env = "^0.6.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
